requirements.txt
tqdm==4.43.0
numpy==1.18.1
opencv_python==4.2.0.32
tensorflow==2.1.0
matplotlib==3.1.3
Pillow==7.0.0
pyTelegramBotAPI==3.6.7
scikit_learn==0.22.2
